+++ src/gfx/effect.hpp
@@ -6,10 +6,10 @@
 
 #include "engine/types.hpp"
 
-void Effect_Feathers(u32 x, u32 y);
-void Effect_Splash(u32 x, u32 y, u8 color);
-void Effect_Explosion(u32 x, u32 y, u8 color);
-void Effect_Smoke(u32 x, u32 y, u8 color);
-void Effect_SmokeClouds(u32 x, u32 y);
-void Effect_Stars(u32 x, u32 y, u8 color);
-void Effect_Destruction(u8 tehoste, u32 x, u32 y);
+void Effect_Feathers(double x, double y);
+void Effect_Splash(double x, double y, u8 color);
+void Effect_Explosion(double x, double y, u8 color);
+void Effect_Smoke(double x, double y, u8 color);
+void Effect_SmokeClouds(double x, double y);
+void Effect_Stars(double x, double y, u8 color);
+void Effect_Destruction(u8 tehoste, double x, double y);
